requirements.txt
fastapi==0.101.0
uvicorn==0.22.0
PyJWT==2.8.0
cryptography==41.0.3
httpx==0.26.0
pytest==7.4.0
coverage==7.2.2
